modules/dagster-ibis/pyproject.toml
[project]
name = "dagster-ibis"
version = "0.1.0"
description = "Base library for interacting with Ibis Tables within dagster."
readme = "README.md"
authors = [
    { name = "James Blackwell", email = "33688964+j-blackwell@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dagster>=1.10.6",
    "ibis-framework>=10.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
